// src/app/api/models/inventory-dashboard-advanced-metrics-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface InventoryDashboardAdvancedMetricsDto {
  averageProductValue: number;
  criticalItemsCount: number;
  inventoryTurnoverRate: number;
  lastMonthMovements: number;
  lowStockCount: number;
  outOfStockCount: number;
  totalActiveProducts: number;
  totalInventoryValue: number;
  totalMovementValue: number;
  totalProducts: number;
}