// src/app/api/models/stock-value-history-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StockValueHistoryDto {
  averageValue: number;
  monthOverMonthGrowth: number;

  /**
   * Period in format YYYY-MM
   */
  period: string;
  productCount: number;
  totalValue: number;
}