// src/app/api/models/product-dimensions-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductDimensionsDto {
  createdAt: string;
  dimensionUnit: 'cm' | 'in' | 'm';
  height: number;
  id: number;
  length: number;
  productId: number;
  updatedAt: string;
  volume?: number;
  volumeUnit: 'ml' | 'l' | 'cm3' | 'm3';
  weight: number;
  weightUnit: 'g' | 'kg' | 'lb' | 'oz';
  width: number;
}