// src/app/api/models/product-packaging-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductPackagingDto {
  barcode?: string;
  baseUnit: string;
  createdAt: string;
  expirationMonths?: number;
  id: number;
  internalSku: string;
  minOrderQuantity: number;
  packageType: string;
  productId: number;
  storageRequirements?: string;
  supplierSku?: string;
  unitsPerPackage: number;
  updatedAt: string;
}