// src/app/api/models/product-inventory-config-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductInventoryConfigDto {
  averageCost: number;
  createdAt: string;
  currentStock: number;
  id: number;
  isActive: boolean;
  lastCost: number;
  maxStockLevel: number;
  minStockLevel: number;
  productId: number;
  reorderPoint: number;
  reservedStock: number;
  supplierId?: number;
  updatedAt: string;
}